OPENNLP-1895: Reject ambiguous matrix source, widen the seed, document quantization, expand tests
Fail loud when a directory holds both model.quantized and model.safetensors
instead of preferring one silently. Widen the QuantizeModel seed argument to
long so the full seed space is expressible. Add a Quantized Models manual
section pointing at the workflow tests. Convert reconstruction and embed-parity
tests to parameterized bit-width cases, add the SentencePiece quantized path
and a both-files-present rejection, and share the cosine and SentencePiece
fixture helpers.

+++ opennlp-extensions/opennlp-embeddings/src/main/java/opennlp/embeddings/cmdline/QuantizeModelParams.java
@@ -47,5 +47,5 @@ interface QuantizeModelParams {
   @ParameterDescription(valueName = "seed",
       description = "the rotation seed; the same matrix, bits, and seed write the same file")
   @OptionalParameter(defaultValue = "0")
-  Integer getSeed();
+  Long getSeed();
 }
